# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Digital Earth API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tiles/cldas/{time_key}/{var}/{z}/{x}/{y}.png": {
      "get": {
        "operationId": "get_cldas_tile_api_v1_tiles_cldas__time_key___var___z___x___y__png_get",
        "parameters": [
          {
            "in": "path",
            "name": "time_key",
            "required": true,
            "schema": {
              "title": "Time Key",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "var",
            "required": true,
            "schema": {
              "title": "Var",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "z",
            "required": true,
            "schema": {
              "maximum": 22,
              "minimum": 0,
              "title": "Z",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "x",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "X",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "y",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Y",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/png": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "PNG tile image"
          },
          "304": {
            "description": "Not Modified"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "Internal Server Error"
          },
          "503": {
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Cldas Tile",
        "tags": [
          "tiles"
        ]
      }
    },
    "/api/v1/tiles/{tile_path}": {
      "get": {
        "operationId": "get_tile_api_v1_tiles__tile_path__get",
        "parameters": [
          {
            "in": "path",
            "name": "tile_path",
            "required": true,
            "schema": {
              "title": "Tile Path",
              "type": "string"
            }
          },
          {
            "description": "Return 302 redirect to object storage when possible",
            "in": "query",
            "name": "redirect",
            "required": false,
            "schema": {
              "default": true,
              "description": "Return 302 redirect to object storage when possible",
              "title": "Redirect",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Tile bytes returned when redirect=false"
          },
          "302": {
            "description": "Redirect to object storage"
          },
          "304": {
            "description": "Not Modified"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Tile",
        "tags": [
          "tiles"
        ]
      }
    },
...
  }
}